Avatar image split up into different parts with randomization.
The avatar image has been split up into 3 different parts. Each part is an image and
each image is rendered on top of each other to form a whole. Each part consists of
a number of images that is randomly selected when the part is rendered.

The root image is rendered at the normal position, while all other parts are rendered
with their margin moved to the left by the size of the image.

Each image currently is a 64x64 png. These images are loaded when the page is
loaded and then a random element is retrieved when the avatar part is rendered.