Context menu for individual chatters with option to remove.
Adds a custom context menu that allows for removing individual chatters for the
duration of the page. The context menu will only open on Chatter elements and when
removed, the ChattersList component will keep a list of removed chatters and will
only render those that have not been removed.

This is a rough first pass of this feature and is setup to allow for more options to be
added in the future.

Close #16.