Implemented chat bubbles to appear when a new message is received.
The Chatters component will now retrieve the message and timestamp of the message for a chatter. This information is passed to the Chatter component.

The Chatter component will keep track of that last time it has rendered a message
based off of the timestamp. If the timestamp is more recent, the Chatter will update
its state so that the new message is displayed in a chat bubble.