Added .env file for client.
The client was created with the create-react-app package, which already contains a
way of reading environment variables from the .env file. However, this behaves
differently than the standard dotenv package installed for a NodeJS server. All
custom environment variables must be prefixed with 'REACT_APP_' in order for those
variables to be placed in the environment.

Create-React-App also recommends committing the .env file and override any local
or other environment changes with .env.local, .env.test, etc. files, which should be
ignored.