Added REACT_APP_TEST_USERS environment variable.
The REACT_APP_TEST_USERS environment variable will bypass the standard fetch
request and will just generate three standard users. This is useful for testing the display of each user in the Chatters component.

The Request function has also been reworked so that parsing of the result is handled
in a separate function. Test data should match the data a live server would send.