Implemented 'api/users/messages' endpoint for chatters.
This commit implements calling the above api on the server. This api call retrieves the
currently available list of messages that have been cached on the server and then
notifies each Chatter component of any new messages they have broadcasted within
the Twitch channel.

The Chatter component has been updated to handle an array of messages and will
continually update the array as it receives new messages from the server. The
component itself will display each message in the list one by one on a 2 second
timer. This behavior is a first pass implementation of displaying multiple chat
messages.