Migrated 'fetch' calls to server from react client project.
This commit migrates over the 'fetch' calls used to retrieve data from the server
about the current state of the connected twitch chat room.

This migration has changed up how the data is retrieved. This new system now uses
async functions along with await function calls instead of the previous 'then'
methods.

A 'fetcher' function is returned which takes in a callback and a specified interval. The
callback should handle data storage and resolve. The 'fetcher' module is only used
for data retrieval.