Implemented 'fetch-store' module to hold data about users.
The 'fetch-store' module holds data about users received from the server. This
module will be responsible for resolving any users that have joined or left the server
since the last time this data was retrieved.

A callback can be provided to listen for events such as adding/removing users from
the client.